Moved cell layout methods to delegate
Created CollectionViewDelegate to manage the cell sizing and data source for the Collection view
