+++ VirtualTourist/Model/CollectionViewDelegate.swift
@@ -1,0 +1,44 @@
+//
+//  CollectionViewDelegate.swift
+//  VirtualTourist
+//
+//  Created by Charlie Scheer on 10/3/19.
+//  Copyright © 2019 Praxsis. All rights reserved.
+//
+
+import Foundation
+import UIKit
+import CoreData
+
+class CollectionViewDelegate: NSObject, UICollectionViewDelegate, UICollectionViewDataSource, UICollectionViewDelegateFlowLayout {
+    var flowLayout: UICollectionViewFlowLayout!
+    
+    init(flowLayout: UICollectionViewFlowLayout) {
+        self.flowLayout = flowLayout
+    }
+    
+    func setupFlowLayoutPreferences() {
+        flowLayout.sectionInset = UIEdgeInsets(top: 10, left: 10, bottom: 10, right: 10)
+        flowLayout.minimumInteritemSpacing = 10
+        flowLayout.minimumLineSpacing = 10
+        flowLayout.itemSize = CGSize(width: 50, height: 50)
+    }
+    
+    func collectionView(_ collectionView: UICollectionView, numberOfItemsInSection section: Int) -> Int {
+        return 5
+    }
+    
+    func collectionView(_ collectionView: UICollectionView, cellForItemAt indexPath: IndexPath) -> UICollectionViewCell {
+        guard let cell = collectionView.dequeueReusableCell(withReuseIdentifier: "CollectionCell", for: indexPath) as? CollectionViewCell else {
+            return UICollectionViewCell()
+        }
+        
+        return cell
+        
+    }
+    
+    func collectionView(_ collectionView: UICollectionView, layout collectionViewLayout: UICollectionViewLayout, sizeForItemAt indexPath: IndexPath) -> CGSize {
+        return CGSize(width: 100, height: 100)
+    }
+    
+}
