+++ VirtualTourist/View/PhotoAlbumViewController.swift
@@ -19,17 +19,20 @@ class PhotoAlbumViewController: UIViewController {
     var mapAnnotation: VirtualTouristMapAnnotation?
     var mapViewDelegate: MapViewDelegate!
     var dataController: DataController!
+    var collectionViewDelegate: CollectionViewDelegate!
     
     
     //MARK: View life cycle Methods
     override func viewDidLoad() {
         super.viewDidLoad()
-        flowLayout.sectionInset = UIEdgeInsets(top: 10, left: 10, bottom: 10, right: 10)
-        flowLayout.minimumInteritemSpacing = 10
-        flowLayout.minimumLineSpacing = 10
-        flowLayout.itemSize = CGSize(width: 50, height: 50)
+        collectionViewDelegate = CollectionViewDelegate(flowLayout: flowLayout)
+        
+        collectionView.delegate = collectionViewDelegate
+        collectionView.dataSource = collectionViewDelegate
         
         setupMapView()
+        
+        
     }
     
     override func viewDidAppear(_ animated: Bool) {
@@ -56,23 +59,4 @@ class PhotoAlbumViewController: UIViewController {
     }
 }
 
-extension PhotoAlbumViewController: UICollectionViewDataSource, UICollectionViewDelegateFlowLayout, UICollectionViewDelegate {
-    func collectionView(_ collectionView: UICollectionView, numberOfItemsInSection section: Int) -> Int {
-        return 5
-    }
-    
-    func collectionView(_ collectionView: UICollectionView, cellForItemAt indexPath: IndexPath) -> UICollectionViewCell {
-        guard let cell = collectionView.dequeueReusableCell(withReuseIdentifier: "CollectionCell", for: indexPath) as? CollectionViewCell else {
-            return UICollectionViewCell()
-        }
-        
-        return cell
-        
-    }
-    
-    func collectionView(_ collectionView: UICollectionView, layout collectionViewLayout: UICollectionViewLayout, sizeForItemAt indexPath: IndexPath) -> CGSize {
-        return CGSize(width: 100, height: 100)
-    }
-    
-}
